renamed signin package for consistency and updated the server string URL in registration and signin view model
