+++ TCSS450Team2Client/app/src/main/res/layout/fragment_sign_in.xml
@@ -5,7 +5,7 @@
     android:id="@+id/layoiut_root"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.auth.sign_in.SignInFragment">
+    tools:context=".ui.auth.signin.SignInFragment">
 
     <EditText
         android:id="@+id/edit_email"
