+++ TCSS450Team2Client/app/src/main/res/values/strings.xml
@@ -1,7 +1,5 @@
 <resources>
     <string name="app_name">TCSS 450 Team 2 Client</string>
-<<<<<<< Updated upstream
-=======
     <string name="hint_first">First Name</string>
     <string name="hint_last">Last Name</string>
     <string name="hint_email">Email</string>
